Use initial touch location for dragstart event
The location of the dragstart event should be the location of the
initial mousedown event, as that's where the drag "originates" from.
Using the touchmove location could be different from the mousedown
location. Indeed, it could even be _outside_ the dragstart element.

The dragenter event still uses the current, touchmove location.